registration and login final

+++ HotelBookingPlatform/src/main/java/com/survival/service/LoginService.java
@@ -6,5 +6,5 @@ import com.survival.entities.Login;
 
 public interface LoginService {
 public Login searchLogin(Login login) throws ClassNotFoundException,SQLException;
-public long searchData(String username) throws ClassNotFoundException,SQLException;
+public int searchData(String username) throws ClassNotFoundException,SQLException;
 }
